+++ app/src/main/java/com/example/nidoqueue/ExpBinomial.java
@@ -1,0 +1,28 @@
+package com.example.nidoqueue;
+
+public class ExpBinomial extends Experiment{
+    private int pass;
+    private int fail;
+
+    public ExpBinomial(User owner) {
+        super(owner);
+        pass = 0;
+        fail = 0;
+    }
+
+    public void increasePass(){
+        pass += 1;
+    }
+
+    public void increaseFail(){
+        fail += 1;
+    }
+
+    public int getPass() {
+        return pass;
+    }
+
+    public int getFail() {
+        return fail;
+    }
+}
