+++ app/src/main/java/com/example/nidoqueue/ForumActivity.java
@@ -1,4 +1,4 @@
 package com.example.nidoqueue;
 
-public class Forum {
+public class ForumActivity {
 }
