+++ app/src/main/java/com/example/nidoqueue/SearchActivity.java
@@ -1,16 +1,15 @@
 package com.example.nidoqueue;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.content.Intent;
 import android.os.Bundle;
-import android.widget.Toast;
 
-public class SignUp extends AppCompatActivity {
+import androidx.appcompat.app.AppCompatActivity;
+
+public class SearchActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.sign_up);
+        setContentView(R.layout.search_trials);
         Intent intent = getIntent();
     }
 }
