+++ app/src/main/res/layout/account_recovery.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#80D9FF"
-    tools:context=".Recovery">
+    tools:context=".RecoveryActivity">
 
     <TextView
         android:id="@+id/code_sent_label"
