+++ app/src/main/res/layout/add_reply.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#80D9FF"
-    tools:context=".Forum">
+    tools:context=".ForumActivity">
 
     <ImageButton
         android:id="@+id/back_button10"
