+++ app/src/main/res/layout/experiment_data.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#80D9FF"
-    tools:context=".ExperimentData">
+    tools:context=".ExperimentDataActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/plots"
