+++ app/src/main/res/layout/forum.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#80D9FF"
-    tools:context=".Forum">
+    tools:context=".ForumActivity">
 
     <Button
         android:id="@+id/add_button2"
