+++ app/src/main/res/layout/search_trials.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     android:background="#80D9FF"
-    tools:context=".Search">
+    tools:context=".SearchActivity">
 
     <ImageButton
         android:id="@+id/search_button2"
