+++ app/src/main/res/layout/sign_up.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#80D9FF"
-    tools:context=".SignUp">
+    tools:context=".SignUpAcitvity">
 
     <TextView
         android:id="@+id/phone_label"
