+++ app/src/main/res/layout/user_profile.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#80D9FF"
-    tools:context=".UserProfile">
+    tools:context=".UserProfileActivity">
 
     <ImageButton
         android:id="@+id/back_button5"
