+++ app/src/main/res/layout/welcome_user.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#80D9FF"
-    tools:context=".SignIn">
+    tools:context=".SignInActivity">
 
     <TextView
         android:id="@+id/created_exp_label"
